requirements.txt
airtable-python-wrapper==0.15.3
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.24.60
botocore==1.27.60
certifi==2022.6.15
charset-normalizer==2.1.1
cloudflare==2.9.12
idna==3.3
jmespath==1.0.1
jsonlines==3.1.0
numpy==1.23.2
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
spin-rewriter-api==1.0.0
urllib3==1.26.12
